rest: move openapi tests to its own file
